+++ flecsi/exec/launch.hh
@@ -5,6 +5,7 @@
 #define FLECSI_EXEC_LAUNCH_HH
 
 #include "flecsi/data/field.hh"
+#include "flecsi/exec/kernel.hh"
 #include "flecsi/exec/task_attributes.hh"
 
 #include <cstddef>
@@ -213,6 +214,12 @@ private:
     return static_cast<D &>(*this);
   }
 };
+
+template<class, class = void>
+struct has_space : std::false_type {};
+template<class S>
+struct has_space<S, std::void_t<typename S::execution_space>> : std::true_type {
+};
 } // namespace detail
 
 template<bool M, class... PP>
@@ -269,8 +276,123 @@ struct launch_domain {
   Color size_;
 };
 
+/// Executor derivation.
+template<class D>
+struct executor_base {
+  /// \see \c executor
+  template<class C, class F>
+  void for_each(C && c, F && f) const {
+    d().named({}).for_each(std::forward<C>(c), std::forward<F>(f));
+  }
+  /// \see \c executor
+  template<class R, class T, class C, class F>
+  [[nodiscard]] T reduce(C && c, F && f) const {
+    return d().named({}).template reduce<R, T>(
+      std::forward<C>(c), std::forward<F>(f));
+  }
+
+private:
+  struct no_label {};
+
+  template<class C>
+  struct forall_t {
+    template<class F>
+    void operator->*(F f) && {
+      e.for_each(std::move(c), std::move(f));
+    }
+    const D & e;
+    C c;
+  };
+  template<class C, class R, class T>
+  struct reduceall_t {
+    template<class F>
+    [[nodiscard]] T operator->*(F f) && {
+      return e.template reduce<R, T>(std::move(c), std::move(f));
+    }
+    const D & e;
+    C c;
+  };
+
+  const D & d() const {
+    return static_cast<const D &>(*this);
+  }
+
+public:
+  template<class P>
+  forall_t<P> flecsi_macro_forall(P && p) const {
+    return {d(), std::forward<P>(p)};
+  }
+  template<class R, class T, class P>
+  reduceall_t<P, R, T>
+  flecsi_macro_reduceall(R *, T *, P && p, no_label) const {
+    return {d(), std::forward<P>(p)};
+  }
+};
+
+// Users need not be aware of these as different classes: they just call any
+// subsequence of {threads, named} to get an object that can launch a kernel.
+// In fact, each produces a class earlier in the following sequence.
+
+/// Parallel operations given a name for debugging or profiling.
+template<class S, unsigned T, unsigned B>
+struct executor : executor_base<executor<S, T, B>> {
+  explicit executor(std::string n) : name(std::move(n)) {}
+
+private:
+  // auto to avoid requiring Kokkos for merely choosing the execution space.
+  auto range(util::id n) const {
+    return exec::policy_type<typename S::execution_space,
+      Kokkos::LaunchBounds<T, B>>(0, n);
+  }
+
+public:
+  /// Apply a function to every element of a range.
+  /// \param c sized random-access range, potentially copied
+  template<class C, class F>
+  void for_each(C && c, F && f) const {
+    kok::parallel_for(
+      name, range(c.size()), std::forward<C>(c), std::forward<F>(f));
+  }
+  /// Reduce the results of a function applied to every element of a range.
+  /// \tparam R reduction operation type
+  /// \tparam A accumulator type
+  template<class R, class A, class C, class F>
+  [[nodiscard]] A reduce(C && c, F && f) const {
+    return kok::parallel_reduce<R, A>(
+      name, range(c.size()), std::forward<C>(c), std::forward<F>(f));
+  }
+
+private:
+  std::string name;
+};
+/// Parallel operations with thread configuration.
+template<class S, unsigned T, unsigned B>
+struct blocks : executor_base<blocks<S, T, B>> {
+  /// Specify a name for an operation.
+  /// \return \c executor
+  auto named(std::string n) const {
+    return executor<S, T, B>(std::move(n));
+  }
+};
+/// A node-local context for potentially parallel operations.
+template<class S>
+struct agent : executor_base<agent<S>> {
+  /// Specify threads and blocks for an operation.
+  /// These are ignored if not supported by the execution space.
+  /// \see \c Kokkos::LaunchBounds
+  /// \return \c blocks
+  template<unsigned T, unsigned B>
+  auto threads() const {
+    return blocks<S, T, B>();
+  }
+  /// \see \c blocks
+  auto named(std::string n) const {
+    return threads<0, 0>().named(std::move(n));
+  }
+};
+
 /// An execution space.
-struct space : data::bind_tag {
+struct space_base : data::bind_tag {
   /// Information about an index launch.
   struct tasks {
     Color size, ///< Number of point tasks launched.
@@ -286,22 +408,45 @@ struct space : data::bind_tag {
   }
 
   template<class T>
-  using keep = std::conditional_t<std::is_base_of_v<space, T>, T, void>;
+  using keep = std::conditional_t<std::is_base_of_v<space_base, T>, T, void>;
 
 private:
   tasks t{};
 };
+/// Execution space operations.
+template<class S>
+struct space : space_base {
+  // Since derived executors should be able to outlive their bases, it makes
+  // sense to allow a base executor to outlive its (potentially copied) space.
+
+  /// Get an executor for potentially parallel operations on this space.
+  agent<S> executor() const {
+    return {};
+  }
+};
+
 /// Single-core execution space.
-struct cpu : space {
+struct cpu : space<cpu> {
   static constexpr processor proc = processor::loc;
+#ifdef KOKKOS_ENABLE_SERIAL
+  using execution_space = Kokkos::Serial;
+#endif // otherwise undefined
 };
 /// GPU execution space.
-struct gpu : space {
+struct gpu : space<gpu> {
   static constexpr processor proc = processor::toc;
+#ifdef KOKKOS_ENABLE_CUDA
+  using execution_space = Kokkos::Cuda;
+#elif defined(KOKKOS_ENABLE_HIP)
+  using execution_space = Kokkos::HIP;
+#endif // otherwise undefined
 };
 /// OpenMP execution space.
-struct omp : space {
+struct omp : space<omp> {
   static constexpr processor proc = processor::omp;
+#ifdef KOKKOS_ENABLE_OPENMP
+  using execution_space = Kokkos::OpenMP;
+#endif // otherwise undefined
 };
 
 template<processor>
@@ -322,6 +467,12 @@ template<>
 struct processor_space<processor::mpi> : processor_space<processor::loc> {};
 template<processor P>
 using processor_space_t = typename processor_space<P>::type;
+
+/// The available accelerated execution space.  Defined as \c gpu or \c omp if
+/// support for one of those is available, otherwise \c cpu.
+using accelerator = std::conditional_t<detail::has_space<gpu>::value,
+  gpu,
+  std::conditional_t<detail::has_space<omp>::value, omp, cpu>>;
 
 // Find the (single) execution space among parameter types (or void):
 template<class T>
@@ -344,7 +495,7 @@ template<class... TT>
 struct param_space<std::tuple<TT...>> {
   using type = typename decltype((
     processor_combine<void>() | ... |
-    processor_combine<space::keep<std::decay_t<TT>>>()))::type;
+    processor_combine<space_base::keep<std::decay_t<TT>>>()))::type;
 };
 
 struct on_t : data::convert_tag {};