+++ tutorial/5-parallel/2-reduceall.cc
@@ -20,25 +20,26 @@ init(canon::accessor<ro> t, field<double>::accessor<wo> p) noexcept {
 } // init
 
 void
-reduce1(canon::accessor<ro> t, field<double>::accessor<ro> p) {
-  auto res = reduceall(c, up, t.cells(), exec::fold::max, double, "reduce1") {
+reduce1(exec::accelerator s,
+  canon::accessor<ro> t,
+  field<double>::accessor<ro> p) noexcept {
+  auto res = s.executor().named("reduce1").reduceall(
+    c, up, t.cells(), exec::fold::max, double) {
     up(p[c]);
   }; // forall
 
   flog_assert(res == 6.0, res << " != 6.0");
-
-} // reduce1
+}
 
 void
-reduce2(canon::accessor<ro> t, field<double>::accessor<ro> p) {
-  auto res = flecsi::exec::parallel_reduce<exec::fold::max, double>(
-    t.cells(),
-    FLECSI_LAMBDA(auto c, auto up) { up(p[c]); },
-    std::string("reduce2"));
+reduce2(exec::accelerator s,
+  canon::accessor<ro> t,
+  field<double>::accessor<ro> p) noexcept {
+  auto res = s.executor().reduce<exec::fold::max, double>(
+    t.cells(), FLECSI_LAMBDA(auto c, auto up) { up(p[c]); });
 
   flog_assert(res == 6.0, res << " != 6.0");
-
-} // reduce2
+}
 
 void
 print(canon::accessor<ro> t, field<double>::accessor<ro> p) noexcept {
@@ -59,8 +60,8 @@ advance(control_policy & p) {
 
   // cpu task, default
   s.execute<init>(canonical, pf);
-  execute<reduce1, default_accelerator>(canonical, pf);
-  execute<reduce2, default_accelerator>(canonical, pf);
+  s.execute<reduce1>(exec::on, canonical, pf);
+  s.execute<reduce2>(exec::on, canonical, pf);
   // cpu_task
   s.execute<print>(canonical, pf);
 }