+++ tutorial/standalone/flaxpy/flaxpy.cc
@@ -88,16 +88,19 @@ const one_field::definition<flaxpy::dist_vector> x_field, y_field;
 
 // Define a task that initializes the elements of the distributed vector.
 void
-initialize_vectors_task(one_field::accessor<flecsi::wo> x_acc,
-  one_field::accessor<flecsi::wo> y_acc) {
+initialize_vectors_task(flecsi::exec::accelerator s,
+  one_field::accessor<flecsi::wo> x_acc,
+  one_field::accessor<flecsi::wo> y_acc) noexcept {
   // Arbitrarily initialize x[i] = i and y[i] = 0.  We use a forall
   // for the latter because it can run in parallel without access to
   // the index variable.
   auto p = x_acc.span().begin();
   for(size_t i :
-    flaxpy::divide_indices_among_colors(flecsi::colors())[flecsi::color()])
+    flaxpy::divide_indices_among_colors(s.launch().size)[s.launch().index])
     *p++ = i;
-  forall(elt, y_acc.span(), "init_y") { elt = 0; };
+  s.executor().forall(elt, y_acc.span()) {
+    elt = 0;
+  };
 }
 
 // Implement an action for the initialize control point.
@@ -107,8 +110,9 @@ initialize_action(flaxpy::control_policy & policy) {
   // Specify one color per process.
   policy.dist_vector_slot.allocate(
     sch, flaxpy::dist_vector::mpi_coloring(sch, flecsi::processes()));
-  flecsi::execute<initialize_vectors_task, flecsi::default_accelerator>(
-    x_field(policy.dist_vector_slot), y_field(policy.dist_vector_slot));
+  sch.execute<initialize_vectors_task>(flecsi::exec::on,
+    x_field(policy.dist_vector_slot),
+    y_field(policy.dist_vector_slot));
 }
 
 // Define a task that assigns Y <- a*X + Y.
@@ -131,13 +135,10 @@ mul_add_action(flaxpy::control_policy & policy) {
 
 // Define a task that adds up all values of Y and returns the sum.
 double
-reduce_y_task(one_field::accessor<flecsi::ro> y_acc) {
-  auto local_sum = reduceall(elt,
-    accum,
-    y_acc.span(),
-    flecsi::exec::fold::sum,
-    double,
-    "reduce_y_task") {
+reduce_y_task(flecsi::exec::accelerator s,
+  one_field::accessor<flecsi::ro> y_acc) noexcept {
+  const auto local_sum = s.executor().reduceall(
+    elt, accum, y_acc.span(), flecsi::exec::fold::sum, double) {
     accum(elt);
   };
   return local_sum;
@@ -146,10 +147,10 @@ reduce_y_task(one_field::accessor<flecsi::ro> y_acc) {
 // Implement an action for the finalize control point.
 void
 finalize_action(flaxpy::control_policy & policy) {
-  double sum = flecsi::reduce<reduce_y_task,
-    flecsi::exec::fold::sum,
-    flecsi::default_accelerator>(y_field(policy.dist_vector_slot))
-                 .get();
+  const double sum = policy.scheduler()
+                       .reduce<reduce_y_task, flecsi::exec::fold::sum>(
+                         flecsi::exec::on, y_field(policy.dist_vector_slot))
+                       .get();
   flog(info) << "The sum over all elements in the final vector is " << sum
              << std::endl;
 }