+++ tutorial/standalone/poisson/app/problem.cc
@@ -11,8 +11,9 @@ using namespace flecsi;
 void
 poisson::action::problem(control_policy & cp) {
   util::annotation::rguard<problem_region> guard;
-  execute<task::eggcarton, default_accelerator>(
-    cp.m, ud(cp.m), fd(cp.m), sd(cp.m), Aud(cp.m));
+  auto & s = cp.scheduler();
+  s.execute<task::eggcarton>(
+    exec::on, cp.m, ud(cp.m), fd(cp.m), sd(cp.m), Aud(cp.m));
   execute<task::io, flecsi::mpi>(exec::on, cp.m, ud(cp.m), "init");
   execute<task::io, flecsi::mpi>(exec::on, cp.m, sd(cp.m), "actual");